update id of table to long and finish MediaDAO test

+++ Lake-Onboarding/src/main/java/com/parabird/uvds/dataLake/onboarding/db/dao/IDAO.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface IDAO<T> {
     void add(T mappingObject);
     void update(T mappingObject);
-    void delete(int id);
-    T getById(int id);
+    void delete(Long id);
+    T getById(Long id);
     List<T> retrieveAll();
 }
